Mails scheduled for confirmation and confirmation endpoint

+++ src/main/java/com/unibuc/boardmania/model/User.java
@@ -27,7 +27,7 @@ public class User {
 
     private String email;
 
-    private Double trustScore;
+    private int trustScore;
 
     @OneToMany(mappedBy = "initiator")
     private List<Event> createdEvents;
